Move application fragments to common folder
