+++ src/graphql/scheduling/fragments/magazine-schedule-list.js
@@ -1,20 +1,20 @@
 import gql from 'graphql-tag';
-import publicationFragment from './magazine-publication';
-import issueFragment from './magazine-issue';
-import sectionFragment from './magazine-section';
+import publicationFragment from '../../common/fragments/magazine-publication';
+import issueFragment from '../../common/fragments/magazine-issue';
+import sectionFragment from '../../common/fragments/magazine-section';
 
 export default gql`
 
 fragment MagazineScheduleList on MagazineSchedule {
   id
   publication {
-    ...MagazinePublicationScheduleList
+    ...CommonMagazinePublication
   }
   issue {
-    ...MagazineIssueScheduleList
+    ...CommonMagazineIssue
   }
   section {
-    ...MagazineSectionScheduleList
+    ...CommonMagazineSection
   }
 }
