+++ src/graphql/scheduling/queries/load-website-options.js
@@ -1,5 +1,5 @@
 import gql from 'graphql-tag';
-import optionFragment from '../fragments/website-option';
+import optionFragment from '../../common/fragments/website-option';
 
 export default gql`
 
@@ -7,7 +7,7 @@ query LoadWebsiteOptions($input: WebsiteOptionsQueryInput!) {
   websiteOptions(input: $input) {
     edges {
       node {
-        ...WebsiteOptionScheduleList
+        ...CommonWebsiteOption
       }
     }
   }
